+++ src/mame/toshiba/t250.cpp
@@ -50,19 +50,18 @@
 
 #include "emu.h"
 
+#include "t250_ccm.h"
+#include "t250_sasi.h"
+
 #include "bus/centronics/ctronics.h"
 #include "bus/nscsi/devices.h"
-#include "bus/rs232/rs232.h"
 #include "cpu/i8085/i8085.h"
 #include "imagedev/floppy.h"
-#include "machine/i8251.h"
 #include "machine/i8257.h"
 #include "machine/i8279.h"
 #include "machine/input_merger.h"
 #include "machine/nscsi_bus.h"
-#include "machine/nscsi_cb.h"
 #include "machine/output_latch.h"
-#include "machine/pit8253.h"
 #include "machine/upd765.h"
 #include "sound/beep.h"
 #include "video/mc6845.h"
@@ -76,681 +75,6 @@
 
 
 namespace {
-
-// 500 ns minimum pulse width for SASI control lines.
-static constexpr attotime SASI_PULSE = attotime::from_nsec(500);
-
-// ============================================================================
-//  t250_sasi_host_device — SASI host adapter (Toshiba bridge board).
-//
-//  Sits on the SASI bus as an external device at refid 7.  Bridges the
-//  8085 CPU's I/O ports 0x00-0x08 onto the SASI control / data lines
-//  using a small state machine that mimics the Intel 8155-based
-//  adapter board on real hardware.  Patrick's review (PR #15372) asked
-//  for this code to live in its own device class subclassing
-//  nscsi_device_interface so that m_scsi_refid is accessible directly,
-//  and so that the controller can be attached to the SASI bus via the
-//  modern set_external_device() idiom.
-//
-//  TODO: Replace the bespoke ACK/RST/SEL emu_timer pulses with an
-//  i8155_device sub-instance whose internal timer drives those edges
-//  on real hardware.  For now the I/O port semantics are preserved
-//  bit-for-bit so the existing CP/M BIOS m32 build (drives C/D/E/F)
-//  continues to work without further changes.
-// ============================================================================
-
-class t250_sasi_host_device : public device_t, public nscsi_device_interface
-{
-public:
-	t250_sasi_host_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock = 0);
-
-	// 8085 I/O port handlers (mapped by the host driver onto 0x00-0x08).
-	u8   hconfg_r();      // port 0x00 read  (REQ bit 1, active low)
-	void hconfg_w(u8 d);  // port 0x00 write (8155 config: accept + ignore)
-	void hdata_w(u8 d);   // port 0x01 write (data out + implicit ACK)
-	u8   hdata1_r();      // port 0x02 read  (data in + implicit ACK)
-	u8   hstat_r();       // port 0x03 read  (buffer-full bit 4, REQ bit 1)
-	u8   hsts_r();        // port 0x08 read  (BSY inverted, C/D, I/O, MSG)
-	void hrset_w(u8 d);   // port 0x08 write (controller reset)
-
-protected:
-	virtual void device_start() override ATTR_COLD;
-	virtual void device_reset() override ATTR_COLD;
-
-	// nscsi_device_interface — invoked whenever any control line on the
-	// bus changes.  We track REQ and BSY edges and dispatch to our
-	// internal handlers, matching the per-signal callbacks the old
-	// nscsi_callback_device wiring used.
-	virtual void scsi_ctrl_changed() override;
-
-private:
-	void sasi_finish_swallow();
-	TIMER_CALLBACK_MEMBER(sasi_ack_off);
-	TIMER_CALLBACK_MEMBER(sasi_sel_off);
-	TIMER_CALLBACK_MEMBER(sasi_rst_off);
-
-	void on_req_change(int state);
-	void on_bsy_change(int state);
-
-	static constexpr u8 SASI_IDLE            = 0;
-	static constexpr u8 SASI_SELECTING       = 1;
-	static constexpr u8 SASI_CMD_FIRST       = 2;
-	static constexpr u8 SASI_PASSTHROUGH     = 3;
-	static constexpr u8 SASI_SWALLOW_CMD     = 4;
-	static constexpr u8 SASI_SWALLOW_GARBAGE = 5;
-	static constexpr u8 SASI_SWALLOW_STATUS  = 6;
-
-	u8 m_sasi_phase = SASI_IDLE;
-	u8 m_sasi_cmd_first = 0;
-	u8 m_sasi_swallow_remaining = 0;
-	bool m_sasi_synth_bsy = false;
-	bool m_sasi_synth_req = false;
-	u32 m_prev_ctrl = 0;            // last seen bus ctrl; edge-detect input
-	emu_timer *m_sasi_ack_timer = nullptr;
-	emu_timer *m_sasi_sel_timer = nullptr;
-	emu_timer *m_sasi_rst_timer = nullptr;
-};
-
-DEFINE_DEVICE_TYPE_PRIVATE(T250_SASI_HOST, t250_sasi_host_device, t250_sasi_host_device, "t250_sasi_host", "Toshiba T-200/T-250 SASI host adapter")
-
-t250_sasi_host_device::t250_sasi_host_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock)
-	: device_t(mconfig, T250_SASI_HOST, tag, owner, clock)
-	, nscsi_device_interface(mconfig, *this)
-{
-}
-
-void t250_sasi_host_device::device_start()
-{
-	save_item(NAME(m_sasi_phase));
-	save_item(NAME(m_sasi_cmd_first));
-	save_item(NAME(m_sasi_swallow_remaining));
-	save_item(NAME(m_sasi_synth_bsy));
-	save_item(NAME(m_sasi_synth_req));
-	save_item(NAME(m_prev_ctrl));
-
-	m_sasi_ack_timer = timer_alloc(FUNC(t250_sasi_host_device::sasi_ack_off), this);
-	m_sasi_sel_timer = timer_alloc(FUNC(t250_sasi_host_device::sasi_sel_off), this);
-	m_sasi_rst_timer = timer_alloc(FUNC(t250_sasi_host_device::sasi_rst_off), this);
-}
-
-void t250_sasi_host_device::device_reset()
-{
-	m_sasi_phase = SASI_IDLE;
-	m_sasi_cmd_first = 0;
-	m_sasi_swallow_remaining = 0;
-	m_sasi_synth_bsy = false;
-	m_sasi_synth_req = false;
-	m_prev_ctrl = 0;
-	m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_ALL);
-	m_scsi_bus->data_w(m_scsi_refid, 0);
-}
-
-void t250_sasi_host_device::scsi_ctrl_changed()
-{
-	const u32 ctrl = m_scsi_bus->ctrl_r();
-	const u32 changed = ctrl ^ m_prev_ctrl;
-	m_prev_ctrl = ctrl;
-	if (changed & S_REQ) on_req_change((ctrl & S_REQ) ? 1 : 0);
-	if (changed & S_BSY) on_bsy_change((ctrl & S_BSY) ? 1 : 0);
-}
-
-TIMER_CALLBACK_MEMBER(t250_sasi_host_device::sasi_ack_off)
-{
-	m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_ACK);
-	// Critical: release the data bus.  Real hardware uses an external buffer
-	// (74LS244 etc.) gated by I/O so the host doesn't drive data while the
-	// target is driving DATA_IN.  Without this release, the last CDB byte
-	// we drove stays on the bus and OR's with any target-driven data byte —
-	// causing every DATA_IN byte the BIOS reads to be corrupted (the asc_sasi
-	// reference implementation handles this via its iio_w callback).
-	m_scsi_bus->data_w(m_scsi_refid, 0);
-}
-
-TIMER_CALLBACK_MEMBER(t250_sasi_host_device::sasi_sel_off)
-{
-	m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_SEL);
-}
-
-TIMER_CALLBACK_MEMBER(t250_sasi_host_device::sasi_rst_off)
-{
-	m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_RST);
-}
-
-// REQ callback from the bus.  In a real SASI/SCSI initiator the host
-// clears ACK when REQ drops.  We rely on the timer-driven ACK pulse so
-// nothing extra is needed here, but the callback keeps the bus library
-// notified that we observe ctrl changes.
-void t250_sasi_host_device::on_req_change(int state)
-{
-	// No-op: ACK is pulsed via m_sasi_ack_timer right after each data
-	// transfer (asc_sasi/bbc_sasi pattern).
-	(void)state;
-}
-
-void t250_sasi_host_device::on_bsy_change(int state)
-{
-	// Target asserted BSY in response to our SEL — selection complete.
-	// Drop SEL and data bus, then advance the bridge state machine into
-	// command phase so the next hdata_w forwards the first CDB byte.
-	// (BIOS polls hsts to detect BSY independently; this callback is what
-	//  drives our state transition, not the BIOS-visible signalling.)
-	if (state && m_sasi_phase == SASI_SELECTING)
-	{
-		m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_SEL);
-		m_scsi_bus->data_w(m_scsi_refid, 0);
-		m_sasi_phase = SASI_CMD_FIRST;
-	}
-}
-
-// Port 0x00 read.  In the BIOS "req" path (not-inch5) this returns the REQ
-// bit at bit 1, active LOW (`ani 02h; rz` returns when bit 1 is zero).
-// Bit 4 ("buffer full") is sampled separately via port 0x03 (hstat_r).
-u8 t250_sasi_host_device::hconfg_r()
-{
-	// bit 1 = REQ (active LOW: 0 = REQ asserted), default 1 (REQ inactive)
-	// bit 4 = buffer-full (active HIGH: 1 = data ready), default 0 (no data)
-	// Start with d = 0xff masked so bit 4 is cleared; set it explicitly when
-	// DTC actually has a byte queued in DATA_IN phase.
-	u8 d = 0xff & ~0x10;
-	const u32 ctrl = m_scsi_bus->ctrl_r();
-	const bool req = (m_sasi_phase == SASI_SWALLOW_CMD ||
-					   m_sasi_phase == SASI_SWALLOW_GARBAGE ||
-					   m_sasi_phase == SASI_SWALLOW_STATUS)
-		? m_sasi_synth_req
-		: bool(ctrl & nscsi_device_interface::S_REQ);
-	if (req)
-		d &= ~0x02;  // bit 1 cleared = REQ asserted
-	// Buffer-full whenever target is driving a byte toward us (S_INP asserted)
-	// and REQ is up.  This covers BOTH DATA_IN (S_CTL=0) and STATUS phase
-	// (S_CTL=1) — BIOS' hgetrs uses req1 to wait for the status byte, just
-	// like hget does for data bytes.  Also synthesized for SWALLOW_STATUS
-	// when we're feeding back the SS_GOOD reply to a swallowed mode-set.
-	if ((m_sasi_phase == SASI_PASSTHROUGH &&
-	     (ctrl & nscsi_device_interface::S_REQ) &&
-	     (ctrl & nscsi_device_interface::S_INP)) ||
-	    (m_sasi_phase == SASI_SWALLOW_STATUS && m_sasi_synth_req))
-		d |= 0x10;
-	return d;
-}
-
-// Port 0x00 write.  Real hardware: 8155 I/O configuration register.  BIOS
-// writes 0x09 once (port A out, port B in, alt 4) before selecting.  We
-// don't model the 8155 — the write is accepted and dropped.
-void t250_sasi_host_device::hconfg_w(u8 data)
-{
-	(void)data;
-}
-
-// Port 0x03 read.  BIOS' "req1" path polls bit 4 ("buffer full", active
-// HIGH) for the first byte of an inbound transfer.  Bit 1 mirrors REQ
-// (active low) for the inch5 hgetrs path.
-u8 t250_sasi_host_device::hstat_r()
-{
-	u8 d = 0;
-	const u32 ctrl = m_scsi_bus->ctrl_r();
-	const bool req_in = (m_sasi_phase == SASI_SWALLOW_STATUS)
-		? m_sasi_synth_req
-		: ((ctrl & nscsi_device_interface::S_REQ) &&
-		   (ctrl & nscsi_device_interface::S_INP));
-	if (req_in)
-		d |= 0x10;        // buffer-full active high
-	// Match hconfg_r: bit 1 = !REQ on the data-in side (BIOS req1 elsewhere
-	// uses port 0x03 only for the buffer-full poll, but keep consistent).
-	const bool req_any = (m_sasi_phase == SASI_SWALLOW_CMD ||
-						   m_sasi_phase == SASI_SWALLOW_GARBAGE ||
-						   m_sasi_phase == SASI_SWALLOW_STATUS)
-		? m_sasi_synth_req
-		: bool(ctrl & nscsi_device_interface::S_REQ);
-	if (!req_any)
-		d |= 0x02;
-	return d;
-}
-
-// Port 0x08 read.  BIOS uses bit 7 as INVERTED BSY (1 = idle, 0 = busy),
-// bit 2 as C/D phase, bit 1 as I/O, bit 0 as MSG.  See bios64.asm
-// hconsel: `rlc; jnc hact` jumps to controller-reset when bit 7 reads 0
-// (meaning a controller is currently busy/hung).
-u8 t250_sasi_host_device::hsts_r()
-{
-	const u32 ctrl = m_scsi_bus->ctrl_r();
-	// Lazy SELECTING -> CMD_FIRST transition.  Callback registration for BSY
-	// changes can fail silently across the slot-card / driver-device boundary
-	// (finder_dummy_tag warnings in -v), so we don't rely on sasi_bsy_w being
-	// called.  BIOS' hsel1 loop polls hsts hundreds of times per microsecond
-	// while waiting for BSY assertion, so doing the transition here is plenty
-	// responsive.
-	if (m_sasi_phase == SASI_SELECTING && (ctrl & nscsi_device_interface::S_BSY))
-	{
-		m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_SEL);
-		m_scsi_bus->data_w(m_scsi_refid, 0);
-		m_sasi_phase = SASI_CMD_FIRST;
-	}
-	u8 d = 0xff;
-	const bool bsy = m_sasi_synth_bsy || bool(ctrl & nscsi_device_interface::S_BSY);
-	if (bsy)
-		d &= ~0x80;       // inverted BSY at bit 7
-	// For pass-through phases, mirror C/D, I/O, MSG; for swallow phases
-	// drive synthetic phase information that matches what BIOS expects.
-	if (m_sasi_phase == SASI_SWALLOW_CMD)
-	{
-		// command phase: C/D high, I/O low, MSG low
-		d |= 0x04;        // bit 2 = C/D
-		// req mirror at bit 1 (also inverted-ish — BIOS' inch5 `req` path
-		// polls hsts bit 1 active low, just like hconfg)
-		if (m_sasi_synth_req)
-			d &= ~0x02;
-	}
-	else if (m_sasi_phase == SASI_SWALLOW_STATUS)
-	{
-		// status phase: C/D high, I/O high, MSG low
-		d |= 0x04;
-		d |= 0x01;        // BIOS hgetrs polls bit 0 implicitly via rlc/jnc
-		if (m_sasi_synth_req)
-			d &= ~0x02;
-	}
-	else
-	{
-		// Pass-through: mirror real bus.  All hsts phase bits are ACTIVE LOW
-		// (same convention as BSY at bit 7).  d starts at 0xff so we need to
-		// CLEAR each bit when the corresponding nscsi line is asserted.
-		//   bit 2 = C/D  (BIOS hget checks `ani 04h; rz` = exit when bit 2=0,
-		//                 i.e. when real C/D goes high = STATUS phase)
-		//   bit 0 = I/O
-		//   bit 1 = MSG
-		if (ctrl & nscsi_device_interface::S_CTL) d &= ~0x04;
-		if (ctrl & nscsi_device_interface::S_INP) d &= ~0x01;
-		if (ctrl & nscsi_device_interface::S_MSG) d &= ~0x02;
-	}
-	return d;
-}
-
-// Port 0x08 write — controller reset.  BIOS pulses this three times in
-// `hact` to recover a hung controller, and once at the start of `hcon` /
-// `hwaitr`.  Pulse the SASI RST line and forcibly return to IDLE.
-void t250_sasi_host_device::hrset_w(u8 data)
-{
-	(void)data;
-	m_scsi_bus->ctrl_w(m_scsi_refid, nscsi_device_interface::S_RST, nscsi_device_interface::S_RST);
-	// Force the target's state machine to observe the RST edge before the
-	// pulse timer can deassert.  Same race as the SEL pulse fix: BIOS' hact
-	// path pulses RST three times in rapid succession to recover a hung
-	// controller, and without this sync any (or all) of them can be missed,
-	// leaving the target in stale state and BIOS spinning in hsel1 forever.
-	machine().scheduler().synchronize();
-	m_sasi_rst_timer->adjust(SASI_PULSE);
-	m_sasi_phase = SASI_IDLE;
-	m_sasi_synth_bsy = false;
-	m_sasi_synth_req = false;
-	m_sasi_swallow_remaining = 0;
-}
-
-// Port 0x01 write — outbound data byte.
-//
-//   IDLE: BIOS' hconsel writes 0xFF to start a selection.  Drive data=0x01
-//         (target id 0) plus SEL; wait for the target to assert BSY.
-//   PASSTHROUGH (cmd or data-out): forward byte to bus and pulse ACK.
-//   CMD_FIRST: cache the first command byte; decide whether to swallow
-//              this command block locally (hcon mode-set) or forward to
-//              the real target.
-//   SWALLOW_CMD: consume payload bytes; on the last byte, hand off to
-//                SWALLOW_STATUS so hgetrs sees an SS_GOOD reply.
-void t250_sasi_host_device::hdata_w(u8 data)
-{
-	switch (m_sasi_phase)
-	{
-	case SASI_IDLE:
-		// BIOS writes 0xFF to select.  Real hardware ANDs the data byte
-		// into a SEL pulse on the bus; we hardwire target id 0.
-		// SEL is pulsed (not held), matching the asc_sasi reference
-		// implementation — SASI targets respond to the SEL edge, not a
-		// continuous level.
-		if (data == 0xff)
-		{
-			m_scsi_bus->data_w(m_scsi_refid, 0x01);                                          // target id 0
-			m_scsi_bus->ctrl_w(m_scsi_refid, nscsi_device_interface::S_SEL, nscsi_device_interface::S_SEL);
-			// Give the nscsi target's state machine a chance to observe the SEL
-			// edge before our timer can deassert it.  Without this, rapid
-			// back-to-back transactions can race: the target's slice doesn't
-			// get scheduled within the 500 ns SEL pulse window, BSY never
-			// asserts, and BIOS' hsel1 loop spins forever.  Confirmed by Dave
-			// 2026-05-23: hang only manifests in normal speed; with -debug on,
-			// the slower scheduler base loop hides the race.
-			machine().scheduler().synchronize();
-			m_sasi_sel_timer->adjust(SASI_PULSE);                             // auto-deassert SEL
-			m_sasi_phase = SASI_SELECTING;
-			// If a real target is at id 0, it will respond by asserting
-			// BSY almost immediately; the next hsts read by BIOS will see
-			// bit 7 drop and proceed to send the first command byte.
-			// If no target is attached, BSY never asserts and BIOS times
-			// out its 256-iteration spin in hsel1, then re-enters hconsel
-			// — which loops forever.  This is fine because the default
-			// configuration has no target attached and the BIOS does not
-			// probe the HD path on a stock boot.
-			if (m_scsi_bus->ctrl_r() & nscsi_device_interface::S_BSY)
-			{
-				// Target picked up — drop SEL, enter CMD_FIRST.
-				m_scsi_bus->ctrl_w(m_scsi_refid, 0, nscsi_device_interface::S_SEL);
-				m_scsi_bus->data_w(m_scsi_refid, 0);
-				m_sasi_phase = SASI_CMD_FIRST;
-			}
-		}
-		break;
-
-	case SASI_SELECTING:
-		// Shouldn't normally happen — BIOS only writes the 0xFF then polls.
-		break;
-
-	case SASI_CMD_FIRST:
-		m_sasi_cmd_first = data;
-		// Recognise the two hcon mode-set opcodes BIOS sends as the very
-		// first byte of a config command block.  Anything else is a real
-		// CDB byte that gets forwarded to the target.
-		//   0xC1 — T-250 DTC-510 drive-type block (6 bytes total)
-		//   0x0C — T-200 ACB-4000 set-parameters block (14 bytes total)
-		// In SASI CDB terms 0xC1 is "LUN=6, opcode=01 (REZERO)" — illegal —
-		// and 0x0C is INITIALIZE DRIVE CHARACTERISTICS, which generic
-		// nscsi_harddisk does not implement.  Forwarding either would cause
-		// the target to reject with CHECK CONDITION and BIOS to loop in
-		// `hcon` forever (no retry limit on this path).  Instead, abort
-		// the target's command phase with a RST pulse and synthesize our
-		// own bus presence so BIOS can drain the payload and read an
-		// SS_GOOD status.
-		if (data == 0xc1 || data == 0x0c)
-		{
-			m_scsi_bus->ctrl_w(m_scsi_refid, nscsi_device_interface::S_RST, nscsi_device_interface::S_RST);
-			m_sasi_rst_timer->adjust(SASI_PULSE);
-			m_sasi_phase = SASI_SWALLOW_CMD;
-			m_sasi_synth_bsy = true;
-			m_sasi_synth_req = true;
-			// 6 more bytes after this one for T-250 (total 6),
-			// 13 more for T-200 (total 14).
-			m_sasi_swallow_remaining = (data == 0xc1) ? 5 : 13;
-		}
-		else
-		{
-			// Real command — forward to the target's CMD phase.
-			m_scsi_bus->data_w(m_scsi_refid, data);
-			m_scsi_bus->ctrl_w(m_scsi_refid, nscsi_device_interface::S_ACK, nscsi_device_interface::S_ACK);
-			m_sasi_ack_timer->adjust(SASI_PULSE);
-			m_sasi_phase = SASI_PASSTHROUGH;
-		}
-		break;
-
-	case SASI_PASSTHROUGH:
-		m_scsi_bus->data_w(m_scsi_refid, data);
-		m_scsi_bus->ctrl_w(m_scsi_refid, nscsi_device_interface::S_ACK, nscsi_device_interface::S_ACK);
-		m_sasi_ack_timer->adjust(SASI_PULSE);
-		break;
-
-	case SASI_SWALLOW_CMD:
-		if (m_sasi_swallow_remaining > 0)
-			m_sasi_swallow_remaining--;
-		if (m_sasi_swallow_remaining == 0)
-			sasi_finish_swallow();
-		break;
-
-	case SASI_SWALLOW_STATUS:
-		// BIOS shouldn't write here in this phase, but be defensive.
-		break;
-	}
-}
-
-void t250_sasi_host_device::sasi_finish_swallow()
-{
-	// All swallowed payload consumed.  BIOS will now do `in hdata1` once
-	// (pickup garbage), then `call hgetrs` which polls REQ via port 0x03
-	// bit 4 and reads the status byte from hdata1.  Move to SWALLOW_GARBAGE
-	// first — that pickup-garbage read returns 0 and advances us to
-	// SWALLOW_STATUS so the subsequent hgetrs sees the SS_GOOD reply.
-	m_sasi_phase = SASI_SWALLOW_GARBAGE;
-	m_sasi_synth_req = false;  // garbage read isn't gated by req1
-}
-
-// Port 0x02 read — inbound data byte.
-u8 t250_sasi_host_device::hdata1_r()
-{
-	u8 data = 0xff;
-	switch (m_sasi_phase)
-	{
-	case SASI_PASSTHROUGH:
-		data = m_scsi_bus->data_r();
-		if (m_scsi_bus->ctrl_r() & nscsi_device_interface::S_REQ)
-		{
-			m_scsi_bus->ctrl_w(m_scsi_refid, nscsi_device_interface::S_ACK, nscsi_device_interface::S_ACK);
-			m_sasi_ack_timer->adjust(SASI_PULSE);
-		}
-		// If the bus just dropped BSY (target completed message-in), return
-		// to idle so the next selection starts clean.
-		if (!(m_scsi_bus->ctrl_r() & nscsi_device_interface::S_BSY))
-			m_sasi_phase = SASI_IDLE;
-		break;
-
-	case SASI_SWALLOW_CMD:
-		// Shouldn't normally happen — BIOS doesn't read hdata1 while still
-		// writing the command block.  Return 0 to stay quiet.
-		data = 0x00;
-		break;
-
-	case SASI_SWALLOW_GARBAGE:
-		// BIOS' `in hdata1; pickup garbage` after the last CDB byte.
-		// Return 0 and advance to SWALLOW_STATUS, asserting synth_req so
-		// hgetrs' req1 poll passes and the next read returns the status.
-		data = 0x00;
-		m_sasi_phase = SASI_SWALLOW_STATUS;
-		m_sasi_synth_req = true;
-		break;
-
-	case SASI_SWALLOW_STATUS:
-		// First read = status byte = SS_GOOD (0x00).
-		// After we deliver it, BIOS' hgetrs runs `in hdata1; in hsts; rlc;
-		// jnc hgw` — looping while REQ is still asserted.  Drop synth_req
-		// and synth_bsy now so the second read sees no REQ and BIOS exits
-		// the wait loop, then return to IDLE.
-		data = 0x00;
-		m_sasi_synth_req = false;
-		m_sasi_synth_bsy = false;
-		m_sasi_phase = SASI_IDLE;
-		break;
-
-	case SASI_IDLE:
-	case SASI_SELECTING:
-	case SASI_CMD_FIRST:
-		// No data to deliver in these phases.
-		break;
-	}
-	return data;
-}
-
-
-static DEVICE_INPUT_DEFAULTS_START(terminal)
-	DEVICE_INPUT_DEFAULTS("RS232_RXBAUD",   0xff, RS232_BAUD_19200)
-	DEVICE_INPUT_DEFAULTS("RS232_TXBAUD",   0xff, RS232_BAUD_19200)
-	DEVICE_INPUT_DEFAULTS("RS232_DATABITS", 0xff, RS232_DATABITS_8)
-	DEVICE_INPUT_DEFAULTS("RS232_PARITY",   0xff, RS232_PARITY_NONE)
-	DEVICE_INPUT_DEFAULTS("RS232_STOPBITS", 0xff, RS232_STOPBITS_1)
-DEVICE_INPUT_DEFAULTS_END
-
-// ============================================================================
-//  t250_ccm_device — Comms Control Module (CCM) board.
-//
-//  Physically a separate PCB carrying the Intel 8251 USART, the Intel 8253
-//  PIT (counter 2 generates baud clock for both TxC and RxC), and the
-//  RS-232 port.  Standard on every T-200/T-250 (Dave confirmed not user-
-//  optional), but lives on its own card edge so a device-class
-//  encapsulation is the right shape.  Patrick's review (PR #15372,
-//  line 1335) asked for this split.
-// ============================================================================
-
-class t250_ccm_device : public device_t
-{
-public:
-	t250_ccm_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock = 0);
-
-	u8   pit_r(offs_t offset)            { return m_pit->read(offset); }
-	void pit_w(offs_t offset, u8 data)   { m_pit->write(offset, data); }
-	u8   usart_r(offs_t offset)          { return m_usart->read(offset); }
-	void usart_w(offs_t offset, u8 data);   // 0xA4/A5 — first-mode-byte quirk
-	void mode_w(u8 data);                   // 0xA6 — CCM mode register
-	u8   cicts_r();                         // 0xA7 — CI/CTS status
-
-	auto rxrdy_handler() { return m_rxrdy_cb.bind(); }
-	auto txrdy_handler() { return m_txrdy_cb.bind(); }
-	bool rxrdy_state() const { return m_rxrdy_state; }
-	bool txrdy_state() const { return m_txrdy_state; }
-	bool cts_r() const       { return m_cts; }
-	bool ci_r()  const       { return m_ri;  }
-
-	// USART input lines forwarded to the CCM's 8251.  These are
-	// invoked by the host's RS232 port callbacks; the inverse paths
-	// (TX/DTR/RTS out of the 8251) are exposed by hooking the host's
-	// m_rs232 directly into our usart().
-	void rxd_w(int state) { m_usart->write_rxd(state); }
-	void dsr_w(int state) { m_usart->write_dsr(state); }
-	void cts_w(int state) { rs232_cts_w(state); }
-	void ri_w(int state)  { rs232_ri_w(state);  }
-	i8251_device &usart() const { return *m_usart; }
-
-protected:
-	virtual void device_start() override ATTR_COLD;
-	virtual void device_reset() override ATTR_COLD;
-	virtual void device_add_mconfig(machine_config &config) override ATTR_COLD;
-
-private:
-	void usart_rxrdy_w(int state);
-	void usart_txrdy_w(int state);
-	void rs232_cts_w(int state);
-	void rs232_ri_w(int state);
-
-	required_device<pit8253_device>    m_pit;
-	required_device<i8251_device>      m_usart;
-
-	devcb_write_line m_rxrdy_cb;
-	devcb_write_line m_txrdy_cb;
-
-	bool m_rxrdy_state = false;
-	bool m_txrdy_state = false;
-	bool m_usart_inited = false;
-	bool m_cts = true;
-	bool m_ri  = false;
-};
-
-DEFINE_DEVICE_TYPE_PRIVATE(T250_CCM, t250_ccm_device, t250_ccm_device, "t250_ccm", "Toshiba T-200/T-250 CCM (Comms Control Module)")
-
-t250_ccm_device::t250_ccm_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock)
-	: device_t(mconfig, T250_CCM, tag, owner, clock)
-	, m_pit(*this, "pit")
-	, m_usart(*this, "usart")
-	, m_rxrdy_cb(*this)
-	, m_txrdy_cb(*this)
-{
-}
-
-void t250_ccm_device::device_start()
-{
-	save_item(NAME(m_rxrdy_state));
-	save_item(NAME(m_txrdy_state));
-	save_item(NAME(m_usart_inited));
-	save_item(NAME(m_cts));
-	save_item(NAME(m_ri));
-}
-
-void t250_ccm_device::device_reset()
-{
-	m_usart->write_cts(0);  // 8251 /CTS tied low on real HW; BIOS handles flow via port 0xA7
-	m_rxrdy_state = false;
-	m_txrdy_state = false;
-	m_usart_inited = false;
-	m_cts = true;
-	m_ri  = false;
-}
-
-void t250_ccm_device::device_add_mconfig(machine_config &config)
-{
-	// CCM-board baud clock is 1.9968 MHz (15.9744 MHz crystal / 8), not the
-	// standard 1.8432 MHz async baud crystal.  Per Dave from the set utility's
-	// divisor table: 19200 -> 6 gives 1996800/6/16 = 20800 (~8% high; lower
-	// rates are exact).
-	PIT8253(config, m_pit, 0);
-	m_pit->set_clk<0>(1'996'800);
-	m_pit->set_clk<1>(1'996'800);
-	m_pit->set_clk<2>(1'996'800);
-	m_pit->out_handler<2>().set(m_usart, FUNC(i8251_device::write_txc));
-	m_pit->out_handler<2>().append(m_usart, FUNC(i8251_device::write_rxc));
-
-	I8251(config, m_usart, 1'843'200);
-	// 8251 outputs route to the host's system-level RS232 port via
-	// parent-tag ("^rs232").  The host owns the RS232 connector so
-	// users can address it as -rs232 X at the command line.
-	m_usart->txd_handler().set("^rs232", FUNC(rs232_port_device::write_txd));
-	m_usart->dtr_handler().set("^rs232", FUNC(rs232_port_device::write_dtr));
-	m_usart->rts_handler().set("^rs232", FUNC(rs232_port_device::write_rts));
-	m_usart->rxrdy_handler().set(FUNC(t250_ccm_device::usart_rxrdy_w));
-	m_usart->txrdy_handler().set(FUNC(t250_ccm_device::usart_txrdy_w));
-	// txd/dtr/rts/rxd/dsr/cts/ri are wired by the host (t250_state)
-	// because the RS232 connector lives at the system level so that
-	// the user can address it as -rs232 X at the command line.
-}
-
-
-void t250_ccm_device::usart_rxrdy_w(int state)
-{
-	m_rxrdy_state = state;
-	m_rxrdy_cb(state);
-}
-
-void t250_ccm_device::usart_txrdy_w(int state)
-{
-	m_txrdy_state = state;
-	m_txrdy_cb(state);
-}
-
-// CCM CI/CTS status register (0xA7).  BIOS reads bit 5 to test /CTS
-// (active low: 0 = ready to send).
-u8 t250_ccm_device::cicts_r()
-{
-	u8 d = 0xff;
-	if (m_cts) d &= ~0x20;
-	return d;
-}
-
-// CCM mode register (0xA6).  BIOS init writes 0x18 then 0x00 around
-// CCM mode register write ("asyn - rxdinh - txdwa" pulse sequence per
-// BIOS handler.lib).  Exact function not modeled — BIOS just pulses
-// 0x18 then 0x00 during init; no observable side-effect on emulated
-// traffic so writes are dropped.
-void t250_ccm_device::mode_w(u8 data)
-{
-}
-
-// MAME RS232 convention: cts_handler is called with state=0 for "CTS asserted"
-// (clear to send) and state=1 for deasserted.  We track the inverted value so
-// m_cts == true means "asserted" (matches cicts_r / ios_r consumers).
-// Same for RI.
-void t250_ccm_device::rs232_cts_w(int state) { m_cts = !state; }
-
-// 8251 write port shim.  BIOS initiu writes a stale A=0x18 to the 8251 data
-// register (port 0xA4) *before* having set up the mode byte.  On real
-// hardware the 8251 ignores CPU writes until properly initialized, but
-// MAME's i8251 accepts the byte and locks TX_READY=0 forever because TxEN
-// is also 0 at that moment and start_tx() never fires.  We drop data-port
-// writes until the BIOS sets the mode byte (any write to 0xA5 after a
-// reset / internal reset).  Control-port writes pass through normally so
-// the 8251 can transition through its mode/cmd init sequence.
-void t250_ccm_device::usart_w(offs_t offset, u8 data)
-{
-	if (offset == 0)  // 0xA4 = data port
-	{
-		if (!m_usart_inited)
-			return;
-	}
-	else  // 0xA5 = control port
-	{
-		m_usart_inited = true;
-	}
-	m_usart->write(offset, data);
-}
-void t250_ccm_device::rs232_ri_w(int state)  { m_ri  = !state; }
-
 
 class t250_state : public driver_device
 {
@@ -772,7 +96,6 @@ public:
 		, m_palette(*this, "palette")
 		, m_rst65_merger(*this, "rst65_merger")
 		, m_ccm(*this, "ccm")
-		, m_rs232(*this, "rs232")
 		, m_rom(*this, "maincpu")
 		, m_dsw(*this, "DSW")
 		, m_cfg(*this, "CFG")
@@ -790,12 +113,12 @@ public:
 	void t250(machine_config &config);
 	void t200(machine_config &config);
 
-private:
-	void common(machine_config &config);
-
+protected:
 	virtual void machine_start() override ATTR_COLD;
 	virtual void machine_reset() override ATTR_COLD;
 
+private:
+	void common(machine_config &config);
 	void mem_map(address_map &map) ATTR_COLD;
 	void io_map(address_map &map) ATTR_COLD;
 
@@ -831,8 +154,6 @@ private:
 	u8 dma_memr(offs_t offset);
 	void dma_memw(offs_t offset, u8 data);
 
-	// SASI hard-disk host adapter (Intel 8155 bridge on real hardware).
-
 	MC6845_UPDATE_ROW(crtc_update_row);
 
 	required_device<i8085a_cpu_device> m_maincpu;
@@ -850,7 +171,6 @@ private:
 	required_device<palette_device> m_palette;
 	required_device<input_merger_device> m_rst65_merger;
 	required_device<t250_ccm_device> m_ccm;
-	required_device<rs232_port_device> m_rs232;
 	required_region_ptr<u8> m_rom;
 	required_ioport m_dsw;
 	required_ioport m_cfg;
@@ -883,23 +203,6 @@ private:
 	bool m_prn_perror = false;
 	bool m_prn_fault  = false;
 	bool m_prn_irq    = false;
-
-
-	// SASI host-adapter state.  The bus + target slot are always
-	// instantiated (so port reads return sensible "idle" values), but with
-	// no hard disk mounted the bus stays quiescent and the BIOS hard-disk
-	// path is never exercised.
-	//
-	// The BIOS' "hcon" routine sends a controller-specific drive-parameters
-	// command block as the very first SASI command (opcode 0xC1 for the
-	// DTC-510-style T-250 setup, opcode 0x0C for the Adaptec ACB-4000-style
-	// T-200 setup).  Generic nscsi_harddisk does not implement these, so we
-	// intercept them locally: pulse the bus to abort the target's command
-	// phase, swallow the remaining payload bytes, then synthesize an
-	// SS_GOOD status to the BIOS.  All subsequent commands (READ_6 0x08,
-	// WRITE_6 0x0A, REQUEST_SENSE 0x03, TEST_UNIT_READY 0x00) pass through
-	// to the bus.
-	// Plain u8 phase tag (enum class doesn't satisfy MAME's save_item is_atom
 };
 
 
@@ -1206,26 +509,6 @@ void t250_state::dma_memw(offs_t offset, u8 data)
 }
 
 
-//**************************************************************************
-//  SASI hard-disk host adapter
-//
-//  Real hardware bridges a SASI bus to the 8085 via an Intel 8155.  We do
-//  not model the 8155 itself — we mimic the five port semantics the BIOS
-//  actually uses.  The bus is always instantiated so reads return sensible
-//  idle values; the default-empty target slot keeps the bus quiescent
-//  unless the user attaches a hard disk image via `-sasi:0 harddisk
-//  -hard1 <path>.chd` (see machine_config below).
-//
-//  All BIOS bit polarities verified against bios64.asm hsel/hconsel/hact/
-//  hcon/hgetrs/hget/hput/req/req1 (lines 1298-1633).  KEY: hsts bit 7 is
-//  INVERTED relative to SASI BSY (1 = idle, 0 = BSY asserted).
-//**************************************************************************
-
-// Pulse widths are not load-bearing on either the BIOS' polling code or
-// the nscsi target; pick something nominal.
-
-
-
 MC6845_UPDATE_ROW(t250_state::crtc_update_row)
 {
 	// Character cell is 10 px wide x 10 scanlines tall (ref manual p.39): 8
@@ -1273,11 +556,11 @@ void t250_state::mem_map(address_map &map)
 	// hardware's bank-switched windows:
 	//
 	//   0x0000-0x0FFF : cold-boot ROM (m_rom_view state 0) vs DRAM
-	//                   (m_rom_view disabled).  Even when ROM is being
-	//                   read, writes pass through to the underlying DRAM
-	//                   chip — that mirrors real hardware where the DRAM
-	//                   is permanently wired and ignores the ROM-enable
-	//                   line for write strobes.
+	//                   (m_rom_view disabled).  The view only overlays the
+	//                   ROM read handler; writes fall through to the base
+	//                   .ram() below, which mirrors real hardware where the
+	//                   DRAM chip is permanently wired and absorbs write
+	//                   strobes regardless of the ROM-enable line.
 	//   0xF800-0xFFFF : video RAM (m_vpg_view state 0) vs character-
 	//                   generator RAM / VPG (state 1).  The underlying
 	//                   DRAM at this range is never visible — one of the
@@ -1285,8 +568,7 @@ void t250_state::mem_map(address_map &map)
 	map(0x0000, 0xffff).ram().share("ram");
 
 	map(0x0000, 0x0fff).view(m_rom_view);
-	m_rom_view[0](0x0000, 0x0fff).rom().region("maincpu", 0)
-		.lw8(NAME([this](offs_t offset, u8 data) { m_ram[offset] = data; }));
+	m_rom_view[0](0x0000, 0x0fff).rom().region("maincpu", 0);
 
 	map(0xf800, 0xffff).view(m_vpg_view);
 	m_vpg_view[0](0xf800, 0xffff).ram().share("vram");
@@ -1299,11 +581,7 @@ void t250_state::io_map(address_map &map)
 	// SASI hard-disk host adapter — own device (see t250_sasi_host_device).
 	// Ports 0x00-0x03 and 0x08 are unused unless a hard disk is attached
 	// via -sasi:0 harddisk.
-	map(0x00, 0x00).rw(m_sasi_host, FUNC(t250_sasi_host_device::hconfg_r), FUNC(t250_sasi_host_device::hconfg_w));
-	map(0x01, 0x01).w(m_sasi_host, FUNC(t250_sasi_host_device::hdata_w));
-	map(0x02, 0x02).r(m_sasi_host, FUNC(t250_sasi_host_device::hdata1_r));
-	map(0x03, 0x03).r(m_sasi_host, FUNC(t250_sasi_host_device::hstat_r));
-	map(0x08, 0x08).rw(m_sasi_host, FUNC(t250_sasi_host_device::hsts_r), FUNC(t250_sasi_host_device::hrset_w));
+	map(0x00, 0x08).m(m_sasi_host, FUNC(t250_sasi_host_device::io_map));
 	map(0x80, 0x88).rw(m_dmac, FUNC(i8257_device::read), FUNC(i8257_device::write));
 	map(0x88, 0x88).w(FUNC(t250_state::dmac_mode_w));   // intercept to trigger mem-to-mem
 	map(0x90, 0x90).rw(m_kbc, FUNC(i8279_device::data_r), FUNC(i8279_device::data_w));
@@ -1586,35 +864,16 @@ void t250_state::common(machine_config &config)
 	m_kbc->in_shift_callback().set(FUNC(t250_state::kbd_ctrl_r));  // shift state -> FIFO bit 6 (BIOS shift)
 	m_kbc->in_ctrl_callback().set(FUNC(t250_state::kbd_shift_r));  // keypad flag -> FIFO bit 7 (BIOS rl8)
 
-	// CCM serial board — own device.  See t250_ccm_device class above.
+	// CCM serial board — own device (carries the 8251 USART, the 8253 baud
+	// PIT, and the RS-232 connector at -ccm:rs232).  See t250_ccm_device.
 	T250_CCM(config, m_ccm);
 	m_ccm->rxrdy_handler().set(m_rst65_merger, FUNC(input_merger_device::in_w<1>));
 	m_ccm->txrdy_handler().set(m_rst65_merger, FUNC(input_merger_device::in_w<2>));
 
-	// RS232 connector lives at the system level so users can address
-	// it as -rs232 X at the command line.  Wire CCM <-> rs232 both ways.
-	RS232_PORT(config, m_rs232, default_rs232_devices, nullptr);
-	m_rs232->set_option_device_input_defaults("terminal",   DEVICE_INPUT_DEFAULTS_NAME(terminal));
-	m_rs232->set_option_device_input_defaults("null_modem", DEVICE_INPUT_DEFAULTS_NAME(terminal));
-	m_rs232->rxd_handler().set(m_ccm, FUNC(t250_ccm_device::rxd_w));
-	m_rs232->dsr_handler().set(m_ccm, FUNC(t250_ccm_device::dsr_w));
-	m_rs232->cts_handler().set(m_ccm, FUNC(t250_ccm_device::cts_w));
-	m_rs232->ri_handler().set(m_ccm, FUNC(t250_ccm_device::ri_w));
-
-	// Default any attached terminal to match the BIOS' default mode
-	// (8 data bits, 1 stop, no parity) at 19200 baud.  Users can override
-	// per-session via -rs232:rxbaud / -rs232:txbaud.
-	// Apply default 19200 8N1 (matching the BIOS' default mode byte 0x6E) to
-	// the rs232 slot devices that have configurable bit-rate ports.  Loopback
-	// has no baud setting (it echoes bits directly).  The "Serial Printer"
-	// device only has RX (host→printer), so it gets a printer-specific defaults
-	// block with no TX baud.
-
 	// SASI bus for the optional hard-disk subsystem.  Slot id 0 is the
 	// target (left empty by default — attach with `-sasi:0 harddisk
-	// -hard1 <path>.chd`).  Slot id 7 is the host adapter (initiator);
-	// NSCSI_CB exposes the raw bus lines for the t250_state h*_r/h*_w
-	// methods that implement the 8155-bridge wire protocol.
+	// -hard1 <path>.chd`).  Slot id 7 is the t250_sasi_host_device, the
+	// 8155-bridge initiator, attached below via set_external_device().
 	// NOTE: ALL 8 sasi:N connectors must be declared (even empty ones),
 	// otherwise refid != SCSI id and ctrl_w(7,...)/data_w(7,...) writes
 	// land in unused bus slots that the target can't see.  Confirmed by
